+++ src/handler.rs
@@ -12,8 +12,8 @@ pub enum HandlerAction {
 }
 
 pub trait EventHandler {
-    fn on_connection(&mut self, client_id: u32, stream: &TcpStream) -> Result<()>;
-    fn on_message(&mut self, client_id: u32, data: &[u8]) -> Result<HandlerAction>;
-    fn on_disconnect(&mut self, client_id: u32) -> Result<()>;
+    fn on_connection(&mut self, client_id: u64, stream: &TcpStream) -> Result<()>;
+    fn on_message(&mut self, client_id: u64, data: &[u8]) -> Result<HandlerAction>;
+    fn on_disconnect(&mut self, client_id: u64) -> Result<()>;
     fn is_data_complete(&mut self, data: &[u8]) -> bool;
 }
